+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # MountsJournal
 
-* Added new options for some classes.
+* Added item durability control.